Move code to examples and change name to collection to make clippy happy
